projects/qt_workspace/scream-jar-server/requirements.txt
flask
flask-cors
bcrypt
click
psycopg2-binary